requirements.txt
pytest~=7.1.2
pep8==1.6.2
marshmallow>=2.15.1
beautifulsoup4==4.4.1
ordereddict==1.1
lxml==4.6.2
six~=1.16.0

requests~=2.27.1
python-edgar~=3.1.3
pandas~=1.4.2
pathlib~=1.0.1
setuptools~=61.2.0